requirements.txt
requests
python-dotenv
tweepy
pytrends
flask
